added progress bar

+++ udv_project/main/views.py
@@ -21,30 +21,53 @@ class TasksListView(APIView):
             raise Http404
         if worker.is_first_day:
             tasks = Task.objects.filter(worker_id=worker.id)[:8]
+            tasks_list = []
+            count_completed_task = 0
+            for task in tasks:
+                if task.is_completed:
+                    count_completed_task += 1
+                tasks_list.append(TasksReadSerializer(task).data)
+            progress = count_completed_task/8
+            return Response({'tasks': tasks_list, 'bar_max': 1.0, 'bar': progress})
+
         else:
             tasks = Task.objects.filter(worker_id=worker.id)
 
-        tasks_list = []
+            task_coef = 1/len(tasks)
+            subtask_coef = 1/20
+            progress = 0
 
-        for task in tasks:
-            subtasks_list = []
-            subtasks = Subtask.objects.filter(task_id=task.id)
-            for subtask in subtasks:
-                subtasks_list.append(SubtaskReadSerializer(subtask).data)
-            task_dict = {'id': task.id,
-                         'worker_id': task.worker_id,
-                         'name': task.name,
-                         'is_completed': task.is_completed,
-                         'subtasks': subtasks_list}
-            tasks_list.append(TasksListSerializer(task_dict).data)
-        return Response({'tasks': tasks_list})
+            tasks_list = []
+            for task in tasks:
+                subtasks_list = []
+                subtasks = Subtask.objects.filter(task_id=task.id)
+
+                if task.is_completed:
+                    progress += task_coef
+                else:
+                    count_completed_subtask = 0
+                    for subtask in subtasks:
+                        if subtask.is_completed:
+                            count_completed_subtask += 1
+                        subtasks_list.append(SubtaskReadSerializer(subtask).data)
+                    progress += count_completed_subtask*subtask_coef*task_coef
+
+                task_dict = {'id': task.id,
+                             'worker_id': task.worker_id,
+                             'name': task.name,
+                             'is_completed': task.is_completed,
+                             'subtasks': subtasks_list}
+                tasks_list.append(TasksListSerializer(task_dict).data)
+            return Response({'tasks': tasks_list, 'bar_max': 1.0, 'bar': progress})
 
     def post(self, request):
         try:
             task = Task.objects.get(id=request.data['task_id'])
         except:
             return HttpResponseBadRequest("We cannot process the request. Not such task")
-        serializer = SubtaskSerializer(data=request.data)
+        request_data = request.data
+        request_data.update({"is_completed": False})
+        serializer = SubtaskSerializer(data=request_data)
         serializer.is_valid(raise_exception=True)
         inst = serializer.save()
         return Response({'post': serializer.data, 'subtask_id': inst.id})
@@ -329,7 +352,9 @@ class TaskView(APIView):
         except:
             return HttpResponseBadRequest("We cannot process the request. Not such worker")
 
-        serializer = TasksSerializer(data=request.data)
+        request_data = request.data
+        request_data.update({"is_completed": False})
+        serializer = TasksSerializer(data=request_data)
         serializer.is_valid(raise_exception=True)
         inst = serializer.save()
         return Response({'post': serializer.data, 'task_id': inst.id})
@@ -354,22 +379,37 @@ class TaskView(APIView):
             instance = Task.objects.get(pk=pk)
         except:
             raise Http404
-        instance.is_completed = bool(request.data['is_completed'])
-        instance.save()
-        worker = Worker.objects.get(id=instance.worker_id)
-        if worker.is_first_day:
-            tasks = Task.objects.filter(worker_id=worker.id)[:8]
-            is_all_completed = True
-            for task in tasks:
-                if not task.is_completed:
-                    is_all_completed = False
-            if is_all_completed:
-                worker.is_first_day = False
+        is_completed_row = bool(request.data['is_completed'])
+        if is_completed_row:
+            worker = Worker.objects.get(id=instance.worker_id)
+            if worker.is_first_day:
+                instance.is_completed = is_completed_row
+                instance.save()
+                tasks = Task.objects.filter(worker_id=worker.id)[:8]
+                is_all_completed = True
                 for task in tasks:
-                    task.delete()
-                worker.save()
-
-        return Response({'post': TasksSerializer(instance).data})
+                    if not task.is_completed:
+                        is_all_completed = False
+                if is_all_completed:
+                    worker.is_first_day = False
+                    for task in tasks:
+                        task.delete()
+                    worker.save()
+            else:
+                subtasks = Subtask.objects.filter(task_id=instance.id)
+                is_subtasks_completed = True
+                for subtask in subtasks:
+                    if not subtask.is_completed:
+                        is_subtasks_completed = False
+                if is_subtasks_completed:
+                    instance.is_completed = is_completed_row
+                    instance.save()
+                else:
+                    return HttpResponseBadRequest("We cannot process the request. Cannot cansel completed task with "
+                                                  "uncompleted subtasks")
+            return Response({'post': TasksSerializer(instance).data})
+        else:
+            return HttpResponseBadRequest("We cannot process the request. Cannot cansel completed task")
 
     def delete(self, request, *args, **kwargs):
         pk = kwargs.get("pk", None)
